Backend/requirements.txt
# Django and database connection
Django==4.2  # Adjust version as needed
mysqlclient==2.2.5

# Validation and error handling
django-validators==1.0.2
drf-yasg==1.21.4  # Optional for API schema and validation

# Testing tools
pytest-django==4.5.2
pytest==7.4.0
coverage==7.2.1

# API tools for checking requests (Optional)
djangorestframework==3.14.0
requests==2.31.0  # For making HTTP requests in tests or handling requests

# Linting and code quality (Optional)
flake8==6.0.0
